Implement send_command_expect
I required a method that did not only use timing to signify when the output is returned from a network device. Using netmiko for 2000+ devices, with a set long delay (as some devices take magnitudes longer to return output) was inefficient.

New method tested on 6500 chassis and 3750x switches.

Happy for anyone to update/change as needed.