Check for username and password, always lowercase
Though checking for 'assword' will capture both 'Passsword' and
'password' it won't capture other forms of interesting capitalisation
(like full caps) of the word. This guarantees that we always check for
the 'password' string in the lowercased output. The same goes for
'username'.

It does make the code a bit less funny to read.