OpenSSH configuration support (specifically for jumpbox proxy support)
This has been implemented by outsourcing paramiko.SSHClient.connect(**kwargs)
to a private method ._return_config()

Not only will this give a more central place to tweak client configuration in
the future, but it serves a nice place to do some conditions and openssh cfg
file reading thanks to paramiko.SSHConfig

As of this commit, ssh_config_path is an __init__ arg defaulting to ~/.ssh/config.
If an unparsable ssh config is passed, Exception will be thrown. We can easily
capture this and error something better if needed, but it's a straightfwd error.

On top of proxycommand support, there is now support for Host Aliasing within
a user SSH config. The only drawback/discussion might need of this is whether
to log the provided hostname or the resolved one (and port).

Username will also be favored from the config if provided under the Host block.