Update cisco_asa enable method
Raise exception if enable password is not accepted. 
Throw an longer warning if enable secret is not set (personally I normally expect the login password to be tried, but instead I thought it would be best to give a longer warning).

I had issues where if the secret was not set, it would throw an issue about not setting base prompt, which wasn't very intuitive.