Fixed condition to check if vdoms are enabled
When connecting to a Fortinet device with vdoms disabled I found that device.vdoms == True. The condition to check was using output.find("Virtual domain configuration: enabled") which evaluated to True even though the result was -1. This could be unique to my setup as I'm using Windows with python 3.4.2, however I thought it would be best to explicitly test for -1.