Added three ansi escape codes.
I stumbled upon these while upgrading the firmware of HP ProCurve/Aruba switches (2530).

The codes I found were:

 - [?6l -> Reset mode screen with option 6: 640 x 200 monochrome (graphics). The number specifies the width of the screen, it can go from 0 to 19, except 7 that resets line wrapping.
 - [?7l -> Reset line wrapping.
 - [2J -> Erase display.

I added to the code the whole range for resetting the width of the screen, the reset of the line wrapping and the erase display. The reset of the line wrapping is a bit redundant but that way it's specified that it is indeed removed.