Adding Fastiron and Netiron telnet support
Added Fastiron and Netiron telnet support.

I had to override several methods because when connected via Telnet both Fastiron and Netiron expect a combination of '\r\n' to be sent at the end of a line. So every method that calls write_channel('\n') hardcoded had to be overridden to send '\r\n'. Might patch base_connection in the future to store the newline character in a variable to allow overriding without the duplication of all the functions. 

Also had to override the 3 config related methods (check_config_mode, config_mode, and exit_config_mode) to set the expected pattern to the equivalent of the prompt (I was inheriting from Cisco Base with truncates this because IOS truncates it).