Update brocade_fastiron_telnet.py
Overriding send_command because of an issue with the whole \r\n nonsense. 

Basically we force sending \r\n because the Brocade requires both for a newline but then the output is sanitized so we only have a \n between each line. So when we go to sanitize the output we remove the length of the command string from the beginning of the output. Because the command string has \r\n on the end of it but the actual output only has \n we were removing 1 extra character.