Bug Fix in SNMP autodetect function
Fixed a small bug in the SNMP autodetect function. There was just a
whitespace issue which caused only the last entry of a particular
vendor in the SNMP_MAPPER_BASE dict to be executed, instead of trying
all of them in the order of priority. My guess is this passed all the
tests because the SNMP_MAPPER_BASE currently only has one entry for each
vendor.