Update fortinet_ssh.py
Several commands in Fortinet needs to be in config mode. The idea here is to propose this feature and be able to pass them. I put by default the config global mode but the user should modify this parameter if wants to enter in the config vdom mode.

I've added also the check_config_mode function & modify consequently the exit_config_mode. 

All those updates it's basically a copy paste of the functions available on Cisco and modified to be relative to fortinet devices.

Sam