make send_config_set available for generic devices
Hi @ktbyers , 

I would like to propose a change on terminal_server. Here is the thing : 
I'm making some changes on Avocent Equipments (console stuff), I don't think that a plugin for it should be necessary and terminal_server is enough to do it. I've been able to pass send_commands with the terminal_server but send_config_set is failing the exit_config_mode. I basically would like to add methods that will do nothing (as for the f5 driver) to allow the user to send config set.

What do you think about this change ?