ssh_autodetect fix for IOS XR
In the SSH_MAPPER_DICT, IOS XR devices have the cmd key variable assigned "show version". This works on some of our IOS XR devices (Version 7.5.2) but not on others (Version 6.1.4[Default]). I think this is due to all the additional output on our 6.1.4 routers which comes to a total of 184 lines. On these versions of XR we can use the command "show version brief" instead which fixes the issue and detects the device as expected.