H3C switches are too slow in executing commands
Compared to Cisco switches, H3C switches are too slow in executing
commands because the value of the pattern is not set in the
check_config_mode function. This function is called twice in both
the config_mode and exit_config_mode, each time using read_channel_timing
instead of read_until_pattern, which greatly extends the duration.
This commit set default pattern of hp_comware to save time when
checking the config mode.