Add support to connect to HPE 1920 and 1950 OfficeConnect switches
HPE 1920 and 1950 switches has standard Comware5 / Comware7 CLI but it needs to be enabled with special command sent to the CLI. This patch add new hpe_officeconnect_model argument to HPComwareBase class. When this argument is populated with hpe1920 or hpe1950 string value, it send magic command immediately after establishing connection with switch and therefore these switches can be handled as standard Comware switches.